+++ collector-media3-exoplayer/src/androidTest/java/com/bitmovin/analytics/media3/exoplayer/PhoneBasicScenariosTest.kt
@@ -78,7 +78,11 @@ class PhoneBasicScenariosTest {
                     player.release()
                 }
             }
+            // wait a bit to make sure the player is released
+            Thread.sleep(100)
             MockedIngress.stopServer()
+            // wait a bit to make sure the server is stopped before next test starts
+            Thread.sleep(100)
         }
 
     @Test