+++ collector-media3-exoplayer/src/androidTest/java/com/bitmovin/analytics/media3/exoplayer/SpecialScenariosTest.kt
@@ -6,6 +6,7 @@ import androidx.media3.exoplayer.ExoPlayer
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
 import com.bitmovin.analytics.api.AnalyticsConfig
+import com.bitmovin.analytics.dtos.EventData
 import com.bitmovin.analytics.enums.StreamFormat
 import com.bitmovin.analytics.media3.exoplayer.api.IMedia3ExoPlayerCollector
 import com.bitmovin.analytics.systemtest.utils.DataVerifier
@@ -52,6 +53,7 @@ class SpecialScenariosTest {
     fun teardown() =
         runBlockingTest {
             MockedIngress.stopServer()
+            Thread.sleep(100)
         }
 
     @Test
@@ -112,12 +114,7 @@ class SpecialScenariosTest {
 
             val progStartup = EventDataUtils.getStartupEvent(progEvents.eventDataList)
             val dashStartup = EventDataUtils.getStartupEvent(dashEvents.eventDataList)
-
-            assertThat(progStartup.streamFormat).isEqualTo(StreamFormat.PROGRESSIVE.toString().lowercase())
-            assertThat(progStartup.progUrl).isEqualTo(TestSources.PROGRESSIVE.progUrl!!.substringBefore("?"))
-
-            assertThat(dashStartup.streamFormat).isEqualTo(StreamFormat.DASH.toString().lowercase())
-            assertThat(dashStartup.mpdUrl).isEqualTo(TestSources.DASH.mpdUrl!!.substringBefore("?"))
+            assertStartupSamples(progStartup, dashStartup)
         }
 
     @Test
@@ -174,11 +171,146 @@ class SpecialScenariosTest {
 
             val progStartup = EventDataUtils.getStartupEvent(progEvents.eventDataList)
             val dashStartup = EventDataUtils.getStartupEvent(dashEvents.eventDataList)
-
-            assertThat(progStartup.streamFormat).isEqualTo(StreamFormat.PROGRESSIVE.toString().lowercase())
-            assertThat(progStartup.progUrl).isEqualTo(TestSources.PROGRESSIVE.progUrl!!.substringBefore("?"))
-
-            assertThat(dashStartup.streamFormat).isEqualTo(StreamFormat.DASH.toString().lowercase())
-            assertThat(dashStartup.mpdUrl).isEqualTo(TestSources.DASH.mpdUrl!!.substringBefore("?"))
+            assertStartupSamples(progStartup, dashStartup)
         }
+
+    @Test
+    fun test_reuseCollectorForMultipleSessions_deactivatePlayWhenReadyBeforePreparingSecondMediaItem() =
+        runBlockingTest {
+            val mediaProgressive = MediaItem.fromUri(TestSources.PROGRESSIVE.progUrl!!)
+            val mediaDASH = MediaItem.fromUri(TestSources.DASH.mpdUrl!!)
+            val collector = IMedia3ExoPlayerCollector.create(appContext, defaultAnalyticsConfig)
+
+            collector.sourceMetadata =
+                metadataGenerator.generate(
+                    title = metadataGenerator.getTestTitle(),
+                    addition = mediaProgressive.toString(),
+                )
+
+            val player1 = ExoPlayer.Builder(appContext).build()
+            withContext(mainScope.coroutineContext) {
+                player1.volume = 0.0f
+                collector.attachPlayer(player1)
+                player1.setMediaItem(mediaProgressive)
+                player1.trackSelectionParameters = forceLowestQuality
+                player1.prepare()
+                player1.play()
+            }
+
+            Media3PlayerPlaybackUtils.waitUntilPlayerHasPlayedToMs(player1, 500)
+
+            withContext(mainScope.coroutineContext) {
+                collector.detachPlayer()
+                // setting a new media while playing, causes implicit autoplay for new media
+                // disabling the autoplay right afterwards, is a scenario we want to test here
+                // since this seemed to cause issues with how linkedin integrated (AN-4778)
+                player1.setMediaItem(mediaDASH)
+                collector.attachPlayer(player1)
+
+                // setting playWhenReady to false, to avoid autoplay
+                player1.playWhenReady = false
+                player1.prepare()
+            }
+
+            // Wait for a bit so that we can make sure this
+            // is not added to the startuptime (and thus verify that we cannot run into AN-4778)
+            Thread.sleep(3000)
+
+            withContext(mainScope.coroutineContext) {
+                player1.play()
+            }
+
+            Media3PlayerPlaybackUtils.waitUntilPlayerHasPlayedToMs(player1, 2000)
+
+            withContext(mainScope.coroutineContext) {
+                collector.detachPlayer()
+                player1.release()
+            }
+
+            Thread.sleep(200)
+
+            val impressions = MockedIngress.waitForRequestsAndExtractImpressions()
+            assertThat(impressions).hasSize(2)
+
+            val progEvents = impressions[0]
+            val dashEvents = impressions[1]
+
+            DataVerifier.verifyHasNoErrorSamples(progEvents)
+            DataVerifier.verifyHasNoErrorSamples(dashEvents)
+
+            val progStartup = EventDataUtils.getStartupEvent(progEvents.eventDataList)
+            val dashStartup = EventDataUtils.getStartupEvent(dashEvents.eventDataList)
+
+            // Since we prepare and wait 5 seconds after preparing, startup should be very fast
+            assertThat(dashStartup.videoStartupTime).isLessThan(500)
+            assertStartupSamples(progStartup, dashStartup)
+        }
+
+    @Test
+    fun test_reuseCollectorForMultipleSessions_setNewMediaItemWhilePlaying() =
+        runBlockingTest {
+            val mediaProgressive = MediaItem.fromUri(TestSources.PROGRESSIVE.progUrl!!)
+            val mediaDASH = MediaItem.fromUri(TestSources.DASH.mpdUrl!!)
+            val collector = IMedia3ExoPlayerCollector.create(appContext, defaultAnalyticsConfig)
+
+            collector.sourceMetadata =
+                metadataGenerator.generate(
+                    title = metadataGenerator.getTestTitle(),
+                    addition = mediaProgressive.toString(),
+                )
+
+            val player1 = ExoPlayer.Builder(appContext).build()
+            withContext(mainScope.coroutineContext) {
+                player1.volume = 0.0f
+                collector.attachPlayer(player1)
+                player1.setMediaItem(mediaProgressive)
+                player1.trackSelectionParameters = forceLowestQuality
+                player1.prepare()
+                player1.play()
+            }
+
+            Media3PlayerPlaybackUtils.waitUntilPlayerHasPlayedToMs(player1, 500)
+
+            withContext(mainScope.coroutineContext) {
+                collector.detachPlayer()
+                // setting a new media while playing, causes implicit autoplay for new media
+                player1.setMediaItem(mediaDASH)
+                collector.attachPlayer(player1)
+                player1.prepare()
+            }
+
+            Media3PlayerPlaybackUtils.waitUntilPlayerHasPlayedToMs(player1, 2000)
+
+            withContext(mainScope.coroutineContext) {
+                collector.detachPlayer()
+                player1.release()
+            }
+
+            Thread.sleep(200)
+
+            val impressions = MockedIngress.waitForRequestsAndExtractImpressions()
+            assertThat(impressions).hasSize(2)
+
+            val progEvents = impressions[0]
+            val dashEvents = impressions[1]
+
+            DataVerifier.verifyHasNoErrorSamples(progEvents)
+            DataVerifier.verifyHasNoErrorSamples(dashEvents)
+
+            val progStartup = EventDataUtils.getStartupEvent(progEvents.eventDataList)
+            val dashStartup = EventDataUtils.getStartupEvent(dashEvents.eventDataList)
+
+            assertStartupSamples(progStartup, dashStartup)
+        }
+
+    private fun assertStartupSamples(
+        progStartup: EventData,
+        dashStartup: EventData,
+    ) {
+        assertThat(progStartup.streamFormat).isEqualTo(StreamFormat.PROGRESSIVE.toString().lowercase())
+        assertThat(progStartup.progUrl).isEqualTo(TestSources.PROGRESSIVE.progUrl!!.substringBefore("?"))
+
+        assertThat(dashStartup.streamFormat).isEqualTo(StreamFormat.DASH.toString().lowercase())
+        assertThat(dashStartup.mpdUrl).isEqualTo(TestSources.DASH.mpdUrl!!.substringBefore("?"))
+    }
 }